Cleanup in mesh_cutting
